Rename util/string.h header to avoid conflicts with the system string.h header
* Depending on the build system, it might decide to add the directories
  of all **/*.h to as -I arguments, which then can cause `#include <string.h>`
  to actually include util/string.h and not the system header.
