+++ bin/templates/src/ast.h.erb
@@ -6,7 +6,7 @@
 #ifndef YARP_AST_H
 #define YARP_AST_H
 
-#include "util/string.h"
+#include "util/yp_string.h"
 #include <stddef.h>
 #include <stdint.h>
