+++ src/util/yp_string.h
